Moved ProfileScreen to profile_screens package
